fix: fix the bug
  - fixed a bug where it was possible to delete locations of other users
  - the method of sending a delete request has been changed

+++ src/main/java/com/pustovalov/weatherapplication/service/LocationService.java
@@ -1,14 +1,17 @@
 package com.pustovalov.weatherapplication.service;
 
-import com.pustovalov.weatherapplication.repository.ILocationRepository;
 import com.pustovalov.weatherapplication.dto.LocationSaveDto;
 import com.pustovalov.weatherapplication.entity.Location;
 import com.pustovalov.weatherapplication.entity.User;
+import com.pustovalov.weatherapplication.exception.ObjectNotFoundException;
+import com.pustovalov.weatherapplication.exception.UnauthorizedLocationAccessException;
+import com.pustovalov.weatherapplication.repository.ILocationRepository;
 import com.pustovalov.weatherapplication.service.mapper.LocationMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @RequiredArgsConstructor
 @Service
@@ -33,10 +36,16 @@ public class LocationService {
         return repository.save(mapper.toEntity(locationSaveDto, user));
     }
 
-    public void delete(Long id) {
+    public void delete(Long id, Long userId) {
         if (id <= 0) {
             throw new IllegalArgumentException("location's id cannot be less than or equal to zero ");
         }
+        Location location = repository.findBy(id).orElseThrow(ObjectNotFoundException::new);
+
+        if ((!Objects.equals(location.getUser().getId(), userId))) {
+            throw new UnauthorizedLocationAccessException();
+        }
+
         repository.delete(id);
     }
 }
